Updated README.md including auto-symlink script
Added more apps to default applications, and listed required packages,
including theme, icons and fonts. Included sample command lines for
installing required packages with `pacman` and `yay`.

Also included a copy-pastedable scriptlet that will first detect any
broken symlinks THAT ARE RELEVANT to this repo, in the users home
directory, and asks whether to delete them. Then, the script will loop
the repo, looking for configs that are missing in the home directory.